Address compiler warnings
1) Signed/unsigned variables. Rationale 1: when comparing int to unsigned int
when possible/sensible, change the unsigned int to a larger, signed int
Rationale 2: for loops over size() use a size_t counter
2) Remove unused variables
3) Limit quick game txs to valid quick game type numbers (0x00)
4) Give default value of 0 to CMappingTX int vars to make switch checks on undefined values impossible

+++ src/betting/bet_v3.cpp
@@ -70,7 +70,7 @@ void GetPLBetPayoutsV3(CBettingsView &bettingsViewCache, const int nNewBlockHeig
                         if (bettingsViewCache.results->Read(ResultKey{leg.nEventId}, res)) {
                             uint32_t betOdds = 0;
                             // if bet placed before 2 mins of event started - refund this bet
-                            if (lockedEvent.nStartTime > 0 && uniBet.betTime > (lockedEvent.nStartTime - Params().BetPlaceTimeoutBlocks())) {
+                            if (lockedEvent.nStartTime > 0 && uniBet.betTime > ((int64_t)lockedEvent.nStartTime - Params().BetPlaceTimeoutBlocks())) {
                                 betOdds = fWagerrProtocolV3 ? refundOdds : 0;
                             }
                             else {
@@ -94,7 +94,7 @@ void GetPLBetPayoutsV3(CBettingsView &bettingsViewCache, const int nNewBlockHeig
                     completedBet = true;
 
                     // if bet placed before 2 mins of event started - refund this bet
-                    if (lockedEvent.nStartTime > 0 && uniBet.betTime > (lockedEvent.nStartTime - Params().BetPlaceTimeoutBlocks())) {
+                    if (lockedEvent.nStartTime > 0 && uniBet.betTime > ((int64_t)lockedEvent.nStartTime - Params().BetPlaceTimeoutBlocks())) {
                         if (fWagerrProtocolV3) {
                             odds = refundOdds;
                         } else {
